added cmake for lib,structurized code into libraries.
