+++ CMakeLists.txt
@@ -4,3 +4,7 @@ set(CMAKE_CXX_STANDART_REQUIRED ON)
 set(CMAKE_TRY_COMPILE_TARGET_TYPE "STATIC_LIBRARY")
 project(gameOfLifeStep3 VERSION 1.0)
 add_executable(gameOfLifeStep3 step3.c)
+add_subdirectory(lib)
+target_include_directories(${PROJECT_NAME} PUBLIC lib)
+target_link_libraries(${PROJECT_NAME} board)
+target_link_libraries(${PROJECT_NAME} readInput)
