+++ lib/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_library(board board.c)
+add_library(readInput readInput.c)
