+++ lib/board.h
@@ -1,0 +1,8 @@
+#ifndef BOARD
+#include <stdio.h>
+#include <stdlib.h>
+#include <unistd.h>
+#define BOARD
+void print(int **matrix,int n,int m);
+void next(int **matrix, int n,int m);
+#endif
