+++ lib/readInput.h
@@ -1,0 +1,8 @@
+#ifndef INPUT
+#include <stdio.h>
+#include <stdlib.h>
+#include <unistd.h>
+#define INPUT
+void read_file_in_matrix(int**matrix);
+
+#endif
